@@ -7,17 +7,17 @@
     "private": true,
     "license": "MIT",
     "dependencies": {
-        "bootstrap": "^4.1.1",
-        "jquery": "^3.3.1",
-        "lockr": "^0.8.4",
-        "query-string": "^5.0.0",
-        "react": "^16.3.2",
-        "react-dom": "^16.3.2",
-        "react-icons": "^2.2.7",
-        "react-loadable": "^5.5.0",
-        "react-router-dom": "^4.2.2",
-        "react-scripts-ts": "^2.16.0",
-        "reactstrap": "^6.0.1"
+        "bootstrap": "5.1.3",
+        "jquery": "3.3.1",
+        "lockr": "0.8.4",
+        "query-string": "7.1.1",
+        "react": "18.0.0",
+        "react-dom": "18.0.0",
+        "react-icons": "4.3.1",
+        "react-loadable": "5.5.0",
+        "react-router-dom": "6.3.0",
+        "react-scripts-ts": "4.0.8",
+        "reactstrap": "9.0.2"
     },
     "scripts": {
         "start": "react-scripts-ts start",
@@ -26,17 +26,28 @@
         "deploy": "ren .\\build\\index.html 200.html & surge .\\build"
     },
     "devDependencies": {
-        "@types/jest": "^22.2.3",
-        "@types/lockr": "^0.8.6",
-        "@types/node": "^10.0.6",
-        "@types/query-string": "^5.1.0",
-        "@types/react": "^16.3.14",
-        "@types/react-dom": "^16.0.5",
-        "@types/react-fontawesome": "^1.6.3",
-        "@types/react-icons": "^2.2.5",
-        "@types/react-loadable": "^5.4.1",
-        "@types/react-router-dom": "^4.2.6",
-        "@types/reactstrap": "^5.0.25",
-        "typescript": "^2.9.1"
+        "@types/lockr": "0.8.6",
+        "@types/node": "17.0.24",
+        "@types/query-string": "6.3.0",
+        "@types/react": "18.0.5",
+        "@types/react-dom": "18.0.1",
+        "@types/react-fontawesome": "1.6.5",
+        "@types/react-icons": "3.0.0",
+        "@types/react-loadable": "5.5.6",
+        "@types/react-router-dom": "5.3.3",
+        "@types/reactstrap": "8.7.2",
+        "typescript": "4.6.3"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
     }
 }
