@@ -1,7 +1,8 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import Lockr from "lockr";
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 import App from "./components/App";
 import registerServiceWorker from "./registerServiceWorker";
 
@@ -12,7 +13,14 @@ declare let localStorageSupport: boolean;
 
 // window.localStorageSupport = typeof(Storage) !== "undefined";
 
-ReactDOM.render(<App />, document.getElementById("root") as HTMLElement);
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+root.render(
+    <React.StrictMode>
+        <BrowserRouter>
+            <App />
+        </BrowserRouter>
+    </React.StrictMode>,
+);
 registerServiceWorker();
 // https://www.typescriptlang.org/docs/handbook/advanced-types.html
 // https://reactstrap.github.io/components/form/
