@@ -1,7 +1,6 @@
 import Lockr from "lockr";
 import QueryString from "query-string";
 import React, { Component } from "react";
-import { match } from "react-router";
 import { IGameInfo, IPlayerInfo, Position } from "../../sharedTypes";
 import MoveTree, { Player } from "../../util/MoveTree";
 import Board from "./Board/Board";
@@ -11,7 +10,7 @@ import GameOverModal from "./GameOverModal";
 
 interface IGameContainerProps {
     location: any;
-    match: match<{ index: number }>;
+    index: string
 }
 
 interface IGameContainerState {
@@ -69,7 +68,7 @@ export default class GameContainer extends Component<
             Lockr.set("saved_games", savedGames.slice(1));
         }
 
-        let index: number = this.props.match.params.index;
+        let index: number = parseInt(this.props.index, 10)
 
         if (isNaN(index) || index < 0 || index >= savedGames.length) {
             index = 0;
